requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pdf2image==1.17.0
pillow==11.1.0
PyMuPDF==1.25.1
PyPDF2==3.0.1
sqlparse==0.5.3
